Split out functions  - Logical separation between file parsing & loading  - Ready for test writing
